+++ ingest/participants/main.go
@@ -198,6 +198,8 @@ func ForOperation(
 			})
 		}
 	case xdr.OperationTypeManageAccountSpecificRule:
+	case xdr.OperationTypeCreateKycRecoveryRequest:
+	case xdr.OperationTypeInitiateKycRecovery:
 	default:
 		err = fmt.Errorf("unknown operation type: %s", op.Body.Type)
 	}